+++ test/astral_test.exs
@@ -4,6 +4,6 @@ defmodule AstralTest do
   doctest Astral
 
   test "returns the package version" do
-    assert Astral.version() == "0.1.3"
+    assert Astral.version() == "0.1.4"
   end
 end